requirements.txt
langchain==0.3.*
langchain-openai==0.2.*
langchain-community==0.3.*
langchain-chroma==0.1.*
langchain-text-splitters==0.3.*
langserve[all]==0.3.*
fastapi==0.115.*
uvicorn[standard]==0.32.*
chromadb==0.5.*
pypdf==4.*
beautifulsoup4==4.*
python-dotenv==1.*
